fix: clarity of filenames, consistent theming, proper indexing for media dropdown, mimebundle rerender sorted mimetypes
