+++ beaker-vue/src/components/sidemenu/SideMenuPanel.vue
@@ -1,5 +1,6 @@
 <template>
     <div class="side-panel" :class="extraClasses" v-if="!lazy || (lazy && loaded)">
+        <div class="side-panel-title">{{ props.label }}</div>
         <div class="side-panel-content">
             <slot></slot>
         </div>
@@ -20,7 +21,7 @@ const props = defineProps([
 ]);
 const loaded = ref<boolean>(!props.lazy);
 
-if (props.noOverflow) {
+if (props.noOverflow !== undefined) {
     extraClasses.value.push('no-overflow')
 }
